Move stack_management_jobs from group3 to own group
